+++ .env
@@ -1,9 +1,14 @@
 FLASK_APP='wsgi.py'
 FLASK_ENV='development'
+
 SECRET_KEY='12345'
+
+TESTING = False
+
 #SQLALCHEMY_DATABASE_URI='sqlite://'
 #SQLALCHEMY_DATABASE_URI='sqlite:///foo2.db'
 SQLALCHEMY_DATABASE_URI='sqlite:///covid-19.db'
+
 WTF_CSRF_SECRET_KEY = '12345'                  # Needed by Flask WTForms to combat cross-site request forgery.
-APP_REPOSITORY='memory' # 'memory' or 'database'
-COVID_INSTALL_PATH='C:\\Users\\iwar006\\Documents\\Python dev\\COVID-19'
+
+COVID_REPOSITORY='database' # 'memory' or 'database'
