+++ config.py
@@ -9,12 +9,15 @@ load_dotenv()
 class Config:
     """Set Flask configuration from .env file."""
 
-    # General Config
-    SECRET_KEY = environ.get('SECRET_KEY')
+    # Flask configuration
     FLASK_APP = environ.get('FLASK_APP')
     FLASK_ENV = environ.get('FLASK_ENV')
 
-    # Database
-    SQLALCHEMY_DATABASE_URI = environ.get("SQLALCHEMY_DATABASE_URI")
+    SECRET_KEY = environ.get('SECRET_KEY')
+
+    # Database configuration
+    SQLALCHEMY_DATABASE_URI = environ.get('SQLALCHEMY_DATABASE_URI')
     SQLALCHEMY_ECHO = False
     SQLALCHEMY_TRACK_MODIFICATIONS = False
+
+    COVID_REPOSITORY = environ.get('COVID_REPOSITORY')
