+++ tests/unit/test_services.py
@@ -1,89 +1,66 @@
-from typing import List
-
-from application.adapters.repository import AbstractRepository, MemoryRepository
-from application.domain.model import User, Article, Tag, make_comment, make_tag_association
-
 from datetime import date
-
-from application.service_layer import services, unit_of_work
 
 import pytest
 
+from application import InMemoryUnitOfWork
+from application.service_layer import services
 from application.service_layer.services import AuthenticationException
 
 
-class FakeUnitOfWork(unit_of_work.AbstractUnitOfWork):
-
-    def __init__(self, repo: MemoryRepository):
-        self.repo = repo
-        self.committed = False
-
-    def commit(self):
-        self.committed = True
-
-    def rollback(self):
-        pass
-
-
-def test_can_add_user(in_memory_repo):
+def test_can_add_user(in_memory_uow):
     new_username = 'jz'
     new_password = 'abcd1A23'
 
-    uow = FakeUnitOfWork(in_memory_repo)
-    services.add_user(new_username, new_password, uow)
+    services.add_user(new_username, new_password, in_memory_uow)
 
-    user_as_dict = services.get_user(new_username, uow)
+    user_as_dict = services.get_user(new_username, in_memory_uow)
     assert user_as_dict['username'] == new_username
 
     # Check that password has been encrypted.
     assert user_as_dict['password'].startswith('pbkdf2:sha256:')
 
 
-def test_cannot_add_user_with_existing_name(in_memory_repo):
+def test_cannot_add_user_with_existing_name(in_memory_uow):
     username = 'thorke'
     password = 'abcd1A23'
 
-    uow = FakeUnitOfWork(in_memory_repo)
+    uow = InMemoryUnitOfWork(in_memory_uow)
     with pytest.raises(services.NameNotUniqueException):
-        services.add_user(username, password, uow)
+        services.add_user(username, password, in_memory_uow)
 
 
-def test_authentication_with_valid_credentials(in_memory_repo):
+def test_authentication_with_valid_credentials(in_memory_uow):
     new_username = 'pmccartney'
     new_password = 'abcd1A23'
 
-    uow = FakeUnitOfWork(in_memory_repo)
-    services.add_user(new_username, new_password, uow)
+    services.add_user(new_username, new_password, in_memory_uow)
 
     try:
-        services.authenticate_user(new_username, new_password, uow)
+        services.authenticate_user(new_username, new_password, in_memory_uow)
     except AuthenticationException:
         assert False
 
 
-def test_authentication_with_invalid_credentials(in_memory_repo):
+def test_authentication_with_invalid_credentials(in_memory_uow):
     new_username = 'pmccartney'
     new_password = 'abcd1A23'
 
-    uow = FakeUnitOfWork(in_memory_repo)
-    services.add_user(new_username, new_password, uow)
+    services.add_user(new_username, new_password, in_memory_uow)
 
     with pytest.raises(services.AuthenticationException):
-        services.authenticate_user(new_username, '0987654321', uow)
+        services.authenticate_user(new_username, '0987654321', in_memory_uow)
 
 
-def test_can_add_comment(in_memory_repo):
+def test_can_add_comment(in_memory_uow):
     article_id = 3
     comment_text = 'The loonies are stripping the supermarkets bare!'
     username = 'fmercury'
 
-    uow = FakeUnitOfWork(in_memory_repo)
-
     # Call the service layer to add the comment.
-    services.add_comment(article_id, comment_text, username, uow)
+    services.add_comment(article_id, comment_text, username, in_memory_uow)
 
     # Retrieve the comments for the article from the repository.
-    comments_as_dict = services.get_comments_for_article(article_id, uow)
+    comments_as_dict = services.get_comments_for_article(article_id, in_memory_uow)
 
     # Check that the comments include a comment with the new comment text.
     assert next(
@@ -91,36 +68,32 @@ def test_can_add_comment(in_memory_repo):
         None) is not None
 
 
-def test_cannot_add_comment_for_non_existent_article(in_memory_repo):
+def test_cannot_add_comment_for_non_existent_article(in_memory_uow):
     article_id = 7
     comment_text = "COVID-19 - what's that?"
     username = 'fmercury'
 
-    uow = FakeUnitOfWork(in_memory_repo)
-
     # Call the service layer to attempt to add the comment.
     with pytest.raises(services.NonExistentArticleException):
-        services.add_comment(article_id, comment_text, username, uow)
+        services.add_comment(article_id, comment_text, username, in_memory_uow)
 
 
-def test_cannot_add_comment_by_unknown_user(in_memory_repo):
+def test_cannot_add_comment_by_unknown_user(in_memory_uow):
     article_id = 3
     comment_text = 'The loonies are stripping the supermarkets bare!'
     username = 'gmichael'
 
-    uow = FakeUnitOfWork(in_memory_repo)
+    uow = InMemoryUnitOfWork(in_memory_uow)
 
     # Call the service layer to attempt to add the comment.
     with pytest.raises(services.UnknownUserException):
-        services.add_comment(article_id, comment_text, username, uow)
+        services.add_comment(article_id, comment_text, username, in_memory_uow)
 
 
-def test_can_get_article(in_memory_repo):
+def test_can_get_article(in_memory_uow):
     article_id = 2
 
-    uow = FakeUnitOfWork(in_memory_repo)
-
-    article_as_dict = services.get_article(article_id, uow)
+    article_as_dict = services.get_article(article_id, in_memory_uow)
 
     assert article_as_dict['id'] == article_id
     assert article_as_dict['date'] == date.fromisoformat('2020-02-29')
@@ -136,37 +109,30 @@ def test_can_get_article(in_memory_repo):
     assert 'Politics' in tag_names
 
 
-def test_cannot_get_article_with_non_existent_id(in_memory_repo):
+def test_cannot_get_article_with_non_existent_id(in_memory_uow):
     article_id = 7
-
-    uow = FakeUnitOfWork(in_memory_repo)
 
     # Call the service layer to attempt to retrieve the Article.
     with pytest.raises(services.NonExistentArticleException):
-        services.get_article(article_id, uow)
+        services.get_article(article_id, in_memory_uow)
 
 
-def test_get_first_article(in_memory_repo):
-    uow = FakeUnitOfWork(in_memory_repo)
-
-    article_as_dict = services.get_first_article(uow)
+def test_get_first_article(in_memory_uow):
+    article_as_dict = services.get_first_article(in_memory_uow)
 
     assert article_as_dict['id'] == 1
 
 
-def test_get_last_article(in_memory_repo):
-    uow = FakeUnitOfWork(in_memory_repo)
-
-    article_as_dict = services.get_last_article(uow)
+def test_get_last_article(in_memory_uow):
+    article_as_dict = services.get_last_article(in_memory_uow)
 
     assert article_as_dict['id'] == 6
 
 
-def test_get_articles_by_date_with_one_date(in_memory_repo):
+def test_get_articles_by_date_with_one_date(in_memory_uow):
     target_date = date.fromisoformat('2020-02-28')
 
-    uow = FakeUnitOfWork(in_memory_repo)
-    articles_as_dict, prev_date, next_date = services.get_articles_by_date(target_date, uow)
+    articles_as_dict, prev_date, next_date = services.get_articles_by_date(target_date, in_memory_uow)
 
     assert len(articles_as_dict) == 1
     assert articles_as_dict[0]['id'] == 1
@@ -175,11 +141,10 @@ def test_get_articles_by_date_with_one_date(in_memory_repo):
     assert next_date == date.fromisoformat('2020-02-29')
 
 
-def _get_articles_by_date_with_multiple_dates(in_memory_repo):
+def _get_articles_by_date_with_multiple_dates(in_memory_uow):
     target_date = date.fromisoformat('2020-03-01')
 
-    uow = FakeUnitOfWork(in_memory_repo)
-    articles_as_dict, prev_date, next_date = services.get_articles_by_date(target_date, uow)
+    articles_as_dict, prev_date, next_date = services.get_articles_by_date(target_date, in_memory_uow)
 
 # Sort out the order on this. Don't rely on order. Try membershis, or geerating a list ansd then ordering it.
     assert len(articles_as_dict) == 3
@@ -191,6 +156,6 @@ def _get_articles_by_date_with_multiple_dates(in_memory_repo):
     assert next_date == date.fromisoformat('2020-03-05')
 
 
-def _get_articles_by_date_with_non_existent_date(in_memory_repo):
-    uow = FakeUnitOfWork(in_memory_repo)
+def _get_articles_by_date_with_non_existent_date(in_memory_uow):
+    pass
     # Should be an empty list.
