+++ wsgi.py
@@ -4,4 +4,5 @@ from application import create_app
 app = create_app()
 
 if __name__ == "__main__":
-    app.run(host='0.0.0.0', port=5000)
+    app.run(host='localhost', port=5000, threaded=False)
+
